数据库连接

+++ mybatis/src/test/java/com/example/test/dao/IUserDao.java
@@ -2,6 +2,6 @@ package com.example.test.dao;
 
 public interface IUserDao {
 
-    String queryUserInfoById(Integer id);
+    String queryUserInfoById(Long id);
 
 }
